+++ R/plot_slopes.R
@@ -67,6 +67,12 @@ plot_slopes <- function(model,
         }
     }
 
+    # order of the first few paragraphs is important
+    # if `newdata` is a call to `typical` or `counterfactual`, insert `model`
+    # should probably not be nested too deeply in the call stack since we eval.parent() (not sure about this)
+    scall <- substitute(newdata)
+    newdata <- sanitize_newdata_call(scall, newdata, model)
+
     valid <- c("dydx", "eyex", "eydx", "dyex")
     checkmate::assert_choice(slope, choices = valid)