+++ inst/tinytest/test-plot_slopes.R
@@ -58,5 +58,27 @@ p <- plot_slopes(mod, variables = "hp", condition = c("wt", "am"))
 expect_snapshot_plot(p, "plot_slopes_two_conditions", tol = 500)
 
 
+# Issue #725: `newdata` argument in plotting functions
+mod <- glm(vs ~ hp + am, mtcars, family = binomial)
+p1 <- plot_slopes(mod, variables = "hp", by = "am", newdata = datagridcf(am = 0:1), draw = FALSE)
+p2 <- avg_slopes(mod, variables = "hp", by = "am", newdata = datagridcf(am = 0:1), draw = FALSE)
+expect_equivalent(p1$estimate, p2$estimate)
+expect_equivalent(p1$conf.low, p2$conf.low, tolerance = 1e-6)
+p3 <- plot_slopes(mod, variables = "hp", by = "am", draw = FALSE)
+p4 <- avg_slopes(mod, variables = "hp", by = "am", draw = FALSE)
+expect_equivalent(p3$estimate, p4$estimate)
+expect_equivalent(p3$conf.low, p4$conf.low)
+expect_true(all(p1$conf.low != p3$conf.low))
+p5 <- plot_slopes(mod, variables = "hp", condition = "am", draw = FALSE)
+p6 <- slopes(mod, variables = "hp", newdata = datagrid(am = 0:1))
+expect_equivalent(p5$estimate, p6$estimate)
+expect_equivalent(p5$conf.low, p6$conf.low)
+expect_true(all(p1$conf.low != p5$conf.low))
+expect_true(all(p3$conf.low != p5$conf.low))
+expect_error(plot_slopes(mod, variables = "hp", condition = "am", by = "am"))
+expect_error(plot_slopes(mod, variables = "hp", newdata = mtcars))
+
+
+
 
 rm(list = ls())